requirements.txt
# AI 视频去字幕功能依赖
opencv-python-headless>=4.8.0
numpy>=1.24.0
pillow>=10.0.0
pytesseract>=0.3.10
